# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CONSTR-SW-2021-2-G5 | T2",
        "description": "Eduardo Soares, Gabriel Fanto, João Brentano e Sofia Arend.",
        "version": "1.0.0"
    },
    "paths": {
        "/turma": {
            "post": {
                "summary": "Registrar uma turma.",
                "tags": ["Turma"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Turma"
                            },
                            "examples": {
                                "turma": {
                                    "value": {
                                        "numero": 31,
                                        "ano": 2021,
                                        "semestre": 2,
                                        "disciplina": "61843c585bb2b615ba46af14",
                                        "horario": "NP"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Turma registrada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "items": {
                                        "$ref": "#/components/schemas/Turma"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado."
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                }
            },
            "get": {
                "summary": "Listar todas as turmas.",
                "tags": ["Turma"],
                "responses": {
                    "200": {
                        "description": "Turmas listadas com sucesso."
                    }
                }
            }
        },
        "/turma/{id}": {
          "delete": {
                "summary": "Deletar turma pelo ID",
                "tags": ["Turma"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted user."
                    },
                    "404": {
                        "description": "No records found."
                    },
                    "403": {
                        "description": "Somente adminstrador pode executar esta operação."
                    }
                }
            },
            "get": {
                "summary": "Listar turmas pelo semestre e horario.",
                "tags": ["Turma"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "value": "6186f7d5a908eaa61845505d"
                    },
                    {
                        "name": "semestre",
                        "in": "query",
                        "required": true,
                        "value": 2
                    },
                    {
                        "name": "horario",
                        "in": "query",
                        "required": true,
                        "value": "NP"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turmas listadas com sucesso."
                    }
                }   
            }  
        },
        "/disciplina": {
            "post": {
                "summary": "Registrar uma disciplina.",
                "tags": ["Disciplina"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Disciplina"
                            },
                            "examples": {
                                "disciplina": {
                                    "value": {
                                        "nome": "Projeto e Arquitetura de Software",
                                        "validade": 2021,
                                        "objetivos": "Aplicar técnicas de reutilização de software.",
                                        "ementa": "Conhecimento das arquiteturas dirigidas por modelos. Desenvolvimento e uso de ferramentas de reutilização de software.",
                                        "codigo": 12,
                                        "creditos": 6,
                                        "cargaHoraria": 80 
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Disciplina registrada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "items": {
                                        "$ref": "#/components/schemas/Disciplina"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado."
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                }
            },
            "get": {
                "summary": "Listar todas as disciplinas.",
                "tags": ["Disciplina"],
                "responses": {
                    "200": {
                        "description": "Disciplinas listadas com sucesso."
                    }
                }
            }
        },
        "/disciplina/{id}": {
            "delete": {
                "summary": "Deletar uma disciplina pelo ID.",
                "tags": ["Disciplina"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Deletar uma disciplina pelo ID.",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Turma deletada com sucesso."
                    },
                    "404": {
                        "description": "No records found."
                    },
                    "403": {
                        "description": "Somente adminstrador pode executar esta operação."
                    }
                }
            },
            "get": {
                "summary": "Listar disciplinas pela carga horaria e validade.",
                "tags": ["Disciplina"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "value": "6186e1606eeafc098dbfa83f"
                    },
                    {
                        "name": "cargaHoraria",
                        "in": "query",
                        "required": true,
                        "value": 60
                    },
                    {
                        "name": "validade",
                        "in": "query",
                        "required": true,
                        "value": 2021
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disciplinas listadas com sucesso."
                    }
                }   
            }
        }
    },
    "components": {
        "schemas": {
            "Disciplina": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "validade": {
                        "type": "string"
                    },
                    "objetivos": {
                        "type": "string"
                    },
                    "ementa": {
                        "type": "string"
                    },
                    "codigo": {
                        "type": "number"
                    },
                    "creditos": {
                        "type": "number"
                    },
                    "cargaHoraria": {
                        "type": "number"
                    }    
                }
            },
            "Turma": {
                "type": "object",
                "properties": {
                    "numero": {
                        "type": "number"
                    },
                    "ano": {
                        "type": "number"
                    },
                    "semestre": {
                        "type": "number"
                    },
                    "disciplina": {
                        "type": "string"
                    },
                    "horario": {
                        "type": "string"
                    }
                }
            }
        }
    }
}